Separated Web-Connector from slug file
